Add PreprocTk::splitContentsBy()
This method will allow us to easily split parser functions and
transclusion arguments when they are represented in PreprocTk form
(or the KV which is the 'contents' array of a PreprocTk token).
This patch only adds the new helper and test cases for it.
It is used in Ibf3fc177713ec487bbace32a4b14b5a103358b5e and
I1c78a6d67e63133bdd338d64e62ca2b19374762a.

Bug: T394834
Change-Id: I7c13a830d596755829b3abc1357f9e9dbd410ef2