Rename {,v3}parserfunction to {old-,}parserfunction
Rename 'parserfunction' to 'old-parserfunction' and rename
'v3parserfunction' to 'parserfunction' in preparation for this
eventually becoming the default.  This patch makes no changes in
external DOM serialization, just in our internal TemplateInfo types.

We also move the assignment of parser function "type" to
TemplateEncapsulator, from DOMRangeBuilder, and update the
documentation of the --v3pf and
$wgParsoidExperimentalParserFunctionOutput options.

Change-Id: Ibde918c3085f645e7ff8bf952707c98c6bcbee64