composer.json: remove --allow-polyfill-parser from `phan` invocation
We added this argument in the ancient days when we were first porting
Parsoid to PHP, and our CI setup was "mostly JavaScript" and we ran
`composer test` from inside of `npm test` and the nodejs CI environment
didn't have the `php-ast` extension installed.

But its been years since that was the case, and the polyfill parser
doesn't support named arguments and causes spurious errors in phan.
Don't mislead devs; require `php-ast` now so phan runs locally the
same way it runs in CI.

Change-Id: Id4fbf004fac792781d705b02b8607b9a3558ab29