ParsoidExtensionAPI: add support for &amp in renderMedia method
When renderMedia method in ParsoidExtensionAPI gets a title
containing '&amp;', it rejects the title as invalid.
This patch allows this behavior with replacement of '&amp;'
with '&'. This change does not change behavior of
the translation itself, only lightens one of the criteria
inside, which caused the rejection.

Added new file to mockDataAccess with name 'File_&_file.jpg'

Bug: T403850
Change-Id: I190c91816ae939a98f96709bef40e945859e5ec7