Separate out "TSR source" from the "template argument frame"
We are using the Frame object to represent two very different things:
on one hand it is the "template argument frame" which stores the
information needed to resolve (eg) `{{{1}}}` in the current context.
On the other hand, it is the "TSR/DSR source" which holds the "source
text" needed to obtain the relevant substring from a TSR or DSR.

In certain contexts we can get TSR/DSR values from "the wrong source"
and it can be hard to debug/diagnose/understand what's going on.
Take one small step towards resolution by storing a pointer to the
source text in the TSR/DSR value itself.  Right now this often uses
Frame::getSource() as a fallback but in the future we can further
separate Source from Frame.

Bug: T405759
Change-Id: I68de0c464eb073254374a7d72b72d721022dd49a