Limit default width to file width
The legacy parser knows the file width while handling the link
parameters and, if no width was provided, it uses the user preference,
possibly scaled by upright, unless the file width is less.  In other
words, the default width is limited to the file width.

However, Parsoid doesn't know the file width before requesting a
transform and can only match it after the fact.  That isn't a problem if
client-side upscaling is used, since we can opt not to do it, as in
01230e08.  If we "must render" though, we're likely rendering the
larger image Parsoid requested.

Here we indicate that a width passed to the data access is as a result
of default sizing so that it can be limited to the file size.

Bug: T407423
Depends-On: I265d3f8630a1c14826d1eecfa50f322662207cef
Change-Id: Ie42c88acf8139d8d25efa1f415c882cae597ac5d