Non extlink in wikilink
Extlink tokenizing skips the Utils::isProtocolValid check in the
presence of directives.  When we eventually bail on the extlink in the
token transform, it's too late to give the closing square bracket back
to form a wikilink.  Regardless though, wikilinks have higher precedence
than extlinks.  Extlinks should be stopped if they're caught stealing a
closing bracket from a wikilink.

Bug: T374445
Bug: T400771
Change-Id: I4c17f343a75e16cd7b509c207bb117d4bd36fe37