Self-closed nowikis are pwrap optional
In f204a33, we inspected the contents of dom fragments if the
representative node was considered a wrapper node, thus didn't need
pwrapping, because that might only be true of the first node of the
fragment.

Here we inspect the contents of all dom fragments, regardless of the
representative node, since that's cheap now that we have rich attributes
and less error prone.

Finally, similar to CleanUp::isEmptyNode, nowiki content should be
inspected to see it's whitespace only, which doesn't need pwrapping.

Bug: T407809
Change-Id: I383662dd793f1e80f42abb214f40f89a8c5fb556