Get the list of behavior switch magic words from core
Previously Parsoid assumed that every magic word starting with __ was
a behavior switch.  This is wrong in two ways: not every localization
of a behavior switch starts with double underscore (this is going to be
deprecated), and some languages (Japanese) use a double-wide
underscore character U+FF3F instead of an ascii underscore.

There were some other bugs uncovered here as well, including the fact
that we used the *magic word ID* of the behavior switch in the
`mw:PageProp/` attribute value, but the `::bswPagePropRegex()` method
was checking against the *underscore-stripped localized values* of the
behavior switch.  This happened to work because `__TOC__` happens to
have the magic word id `toc` (for example), but was not correct.

SiteConfig::bswRegexp() has been removed; implementations are no longer
required to implement this method.  Instead, the new method
SiteConfig::getDoubleUnderscoreIDs() will provide the magic word IDs
for the behavior switches.

SiteConfig::getMagicWordForVariable() was also handling case-sensitivity
inconsistently with the other ::getMagicWordFor*() method; that has
been fixed.

Updated all .reqdata and baseconfig files used for testing to
incorporate the new API responses from
I85e1a50451d65626e55f1f83c8f05d42c2658d54; in the process a few test
cases changed and Parsoid recognizes a few new parser functions added
since 2024.

Bug: T407289
Bug: T407290
Depends-On: I85e1a50451d65626e55f1f83c8f05d42c2658d54
Change-Id: Ia8b52d4f03e0df3b29d3a845b045c3f244acec0c