Limit transferring typeof to mw:Transclusion
Typeof can be multivalued and transferring the entire thing isn't
always appropriate.

The patch in I2476dd139dd401134bbe9a8581e21105f81fc1fe changed the
definition of rendering transparent nodes to include empty extension
tags that are given a placeholder link.  However, prior to that patch,
the same issue seems like it would arise if an extension tag returned
only a rendering transparent node.

The crasher from rttesting on
metawiki:Stewards/Elections_2016/Statements/Syum90
was because the extension goes through tree building as an
mw:DOMFragment and that typeof was being copied to the div it was being
migrated into.  Then, when we get to unpacking, none of the data for
that type of node is available on the div.

A test is added which crashes without this patch.

Another test is added to show that inappropriate data is being copied to
the new range start node and not being remove from old range start node.

Bug: T370751
Bug: T378906
Follows-Up: I468ae282e20252f09d5c642890e58c272d6c3ad6
Change-Id: I97933437d4177f87b4c1e758f02b7ab2548a0c61