Recognize dash-prefixed Parsoid fragment markers
In I5fe1a87d86fd68ec0b07a03e29564099de1ebb8c we moved the code which
recognized parsoid fragment markers from template_preproc to
tplarg_or_template, but missed the fact that template_preproc is
*also* called (indirectly) from the `lang_variant_or_tpl` rule,
where it ensures that `-{{...}}` is parsed as a template *not*
language variant markup, in accordance with the general
rightmost-precedence of delimiter parsing in the preprocessor.

This could be fixed either by moving `parsoid_fragment_marker` into
the `template` rule (which is invoked eventually by both
`tplarg_or_template` and `lang_variant_or_tpl`), or by adding
`parsoid_fragment_marker` to `lang_variant_or_tpl`.  I've chosen the
former.  In the process, it turned out that some of the lookahead
rules needed to ensure proper rightmost precedence of braces also
needed to accomodate the fact that fragment markers essentially have
the highest precedence: `{{{#parsoid-fragment...}}}` should parse a
single `{` followed by a fragment marker, not as a template argument.
This was pretty straightforward to fix in the preproc_piece rules; the
backport to the older form of the lookahead used in tplarg_or_template
and lang_variant_or_tpl was slightly messier but mostly a mechanical
substition.

Tests added to confirm that priorities are handled correctly between
sequences of braces and parsoid fragment markers.

Bug: T408391
Follows-Up: I5fe1a87d86fd68ec0b07a03e29564099de1ebb8c
Change-Id: I5f3a2f314790c37df09211d3336969afbd59d9a8