DataMwAttrib: Use live DocumentFragment for html values
This avoids the need for serialization/deserialization, especially
when visiting attributes during traversal.

The clarified types pointed out an inconsistent use of DataMwAttribs
during rich attribute serialization for "attributes with special
semantics" (aka, attributes not starting with `data-`). We now use
a `{"rich":[....]}` wrapper to encode these to ensure they are not
confused for the existing `txt` and `html` type of attribute values.

Rich attributes are only used in production for data- attributes,
so this new wrapper won't affect production.

Change-Id: I8982254cb4eb821797bbe9d675927189f726aea6