Span wrap fostered fragments earlier
Pwrapping of DOMFragments was changed in
I383662dd793f1e80f42abb214f40f89a8c5fb556.

Prior to that patch, DOMFragments were considered pwrap-optional only if
their content had been span wrapped (the first node, really) and their
content itself was all pwrap-optional.

After the patch, DOMFragments were simply considered pwrap-optional if
all their content (without needing the span wrapping) was
pwrap-optional.

The change meant more content avoided pwrapping, fixing T407809.

However, the patch changed PWrap::pWrapOptional to achieve the above
and that method is also used in MarkFosteredContent.  Prior to the
patch, a nested pfragment consisting entirely of a comment would have
been span wrapped and the fostered flag placed on the span.  Now, the
flag is placed on the DOMFragment and transferred to the fragment
children during unpacking.

We already had code to ensure the fragment would be span wrapped if
needed when transferring the fostered flag, however, it was done after
updating dsr where the crash occurred.  Updating dsr assumed the first
node of the fragment was necessarily an element, which it should be in
the three conditions, for now.

The pages had pfragments consisting entirely of a comment because we'd
reached "WARNING: template omitted, post-expand include size too large"

Fixes the crashers on
https://fi.wikipedia.org/wiki/Wikiprojekti:Wiki_Loves_Monuments/Puuttuvat_paikkatiedot_2?useparsoid=1
https://bn.wikivoyage.org/wiki/ইয়েলোস্টোন_ন্যাশনাল_পার্ক?useparsoid=1

Bug: T408958
Change-Id: I4534808d9381a4f75e31e1953e9b12461c3b5a08