DOMTraverser: Allow invoking handlers *before* attribute-embedded HTML
This is needed for processing language converter markup, where we want
to prevent recursion into the data-mw-variant attributes, but allow
recursion into other embedded attributes.  We do this by matching
language converter markup and skipping the node contents before
continuing the traversal into the attributes and children of the
node.

Change-Id: Ib095bc9475171d0f78095af7fabd509e75074341