Ensure that we can round-trip through "embedded page bundle" format
Ordinary PageBundle objects use a JsonCodec for encoding/decoding so
arrays and objects within the `parsoid` key will stay distinct. But
in the "embedded page bundle" format we were using a naive json
encode/decode that would result in turning all JSON objects {} into
JSON arrays [] when we tried to round-trip.

Use a proper JsonCodec to ensure that we can round-trip cleanly.
This is compatible with existing content (if there is any) serialized
with the prior embedded page bundle format: the objects would turn
into arrays, but that's what was happening prior to this patch anyway.

Change-Id: I4214f60582d2315ab86555fe61eab91d210b80d9